covscan.spec: provide all configuration alternatives
... in a single build of a source RPM package.  As Covscan was packaged
in 2014 (commit 6c8a0eb1af9efb7752189eb201ebe074bf09b132), it was
necessary to build RPM packages separately for production and stage
instances of Covscan.

Thanks to the RemovePathPostfixes tag of RPM, we can build all the
configuration (devel, stage, prod) alternatives in a single build.
The instance-specific configuration is moved to separate subpackages
that conflict with each other and their dependencies make sure that
exactly one of them is installed.