docker-compose: add build options and profiles
This makes it possible to use just `compose up` in order to build and
prepare composition.

We also add profiles to distinguish usage between unneeded containers in
a `osh`-able system and the rest.

This integrates changes to scripts as well, thus we refactor them to
take profiles into account and add the `--no-interactive` flag to the
`deploy.sh` script, which should run the client container by default on
MacOS.

Signed-off-by: Isabella Basso do Amaral <idoamara@redhat.com>