gitlab-ci: fix `GITLAB_API_TOKEN` sourcing
As per `gitlab-job-guard` docs [1], the `--private-token` flag should
work, but taking a look at its argument parser [2] we see that this is
not the case, as only `-t`, `--access-token` and `--api-token` should
work.

Also, this variable should be provided by the user and masked, thus
apply `base64` decoding before sourcing it to `gitlab-job-guard`.

[1] - https://gitlab.com/shalomb/gitlab-job-guard
[2] - https://gitlab.com/shalomb/gitlab-job-guard/-/blob/develop/gitlab-job-guard/gitlab-job-guard.py#L222

Signed-off-by: Isabella Basso do Amaral <idoamara@redhat.com>