covscanhub/static: drop from the git repo
As pointed out by Isabella, the `static` directory should actually
be created by Django itself during installation of the application(s).

Some of the files are copied from the `media` directory and other
files are copied from Django/Kobo.  Maintaining an outdated copy
of the `static` directory in the git repo is not useful.

Fixes: https://gitlab.cee.redhat.com/covscan/covscan/-/issues/145
Closes: https://gitlab.cee.redhat.com/covscan/covscan/-/merge_requests/159