client/commands/shortcuts: simplify verify_brew_koji_build
Previously, the code used a simple idea to check existence of the given build
in a rather involved way.  If the dist tag contains the 'fc' substring,
first check Koji and then Brew.  Otherwise, start with Brew first.

This commit retains this logic but implements it in a much simpler way.