packaging: make it possible to build covscan-client for EPEL-9
Since commit 1f757da1871625c566dbbf7379394d831474b2ee we require most
of the run-time dependencies in the buildroot.  Some of them are not
yet available in EPEL-9, which caused the whole build to fail.  This
workaround makes it possible to build at least functional up2date client
for EPEL-9.