common/utils/conf: remove get_home_dir function
It was used only in this module and it can be safely replaced with a
single call to os.path.expanduser('~') because a system where $HOME
points to some random non-existent path is essentially broken.

Nevertheless, osh-cli can be used on such systems as long as the
user has sufficient permissions to create the whole path that $HOME
points to.