hub: remove jQuery from templates
The features of jQuery can be easily substituted by modern JavaScript
so there is no need to use it any more.  Moreover, we are bundling a very
old version of this library.

Unfortunately, the sourcing of the jQuery script in stat/details.html
cannot be removed yet since Morris.js still depends on it.

Related: https://gitlab.cee.redhat.com/covscan/covscan/-/issues/192