containers: scripts/init-db: fix db restore
We need to start the hub so it configures the db for us. But we can't
(and shouldn't) use Django's `loaddata` because of the format of our
data, and because it is slower than restoring directly, thus we need to
stop the server. We also need to reset the db while keeping permissions,
so we drop it and recreate it with the default user (set in
compose.yaml).

Notice that if we hadn't started the hub we also wouldn't be able to
access the db without overriding the defaults.

Fixes: dc9a8e2d ("containers/scripts: merge database restore into init-db script")

Signed-off-by: Isabella do Amaral <idoamara@redhat.com>