packaging: do not include `static` directory into SRPM
Since commit 1f757da1871625c566dbbf7379394d831474b2ee, we do not
maintain the `static` directory in the git repo.  However, if the
directory was created in the working directory (for example through a
bind mount from the development hub container), the directory was by
mistake included in the source RPM package generated by `make srpm`.
This resulted in broken Django Admin UI in case we build a SRPM from
Fedora and installed the resulting packages on a RHEL-8 system.

This commit also removes the `osh/hub/fixtures` directory from the
list because it does not exist in the git repository any more.

Fixes: commit 1f757da1871625c566dbbf7379394d831474b2ee