worker: run osh-worker in foreground in the systemd unit
As pointed out by Lukas Nykryn, the double-forking code provided
by kobo-worker is not implemented in the way that systemd expects
(documented in the daemon(7) man page) and causes the service manager
to misbehave.  Until we fix the compatibility issues in kobo-worker,
it is better to run the daemon in the foreground mode.

As the `LOG_FILE` field in the `worker.conf` configuration file seems
to be ignored in the foreground mode, we configure output redirection
in the unit file.