hub: fix `ScanBindingMixin.latest_packages_scans()`
Previously, the method was excessively slow and complicated.  Moreover,
it always listed only the newest `ScanBinding` with given base NVR which
is not correct.  The given NVR may be a base for more scans in different
releases, e.g. for a build in z-stream and next y-stream.  The old query
would disregard the one that is older.

Successful scans that do not have parents set are the latest for the
given release.  Therefore, it is sufficient to check only this property.

Fixes: f719a64d8534110830ff9b76acad78ab4ce93f7e ("latest runs: return list of runs per base")
Fixes: 25b0506f4c227830a01f3a504d0c12dcc2eed8a6 ("Add searching in latest runs")
Resolves: https://gitlab.cee.redhat.com/covscan/covscan/-/issues/237