hub: implement the listing of overdue scans as a query
Previously, the code used a list comprehension to get the IDs of given
scan bindings, thus, the whole relation had to be searched which can be
slow and expensive.  Also, it cannot be further optimised by Django or
the underlying database engine.

This commit implements the listing of overdue scans as a proper database query.

Fixes: bdf952a8511eabb6d6176cb4071a7f3de34caaeb ("Filter and search by overdue runs.")
Resolves: https://gitlab.cee.redhat.com/covscan/covscan/-/issues/237