hub: Add testing module
Add a Python module that extends Django's `TestCase` and implements
tools for populating the database with test data and analyzing the
content of a response returned by a client.

The purpose of this module is to provide a place for a code shared
and used by unit tests.

The initial set of tools provided by this module are used by unit
tests that test the admin interface displaying scans.

Signed-off-by: Jiří Kučera <jkucera@redhat.com>