packaging: do not restart osh-worker upon update
osh-worker does not implement `reload` and `restart` would interrupt
running tasks.  Until a proper `reload` operation is implemented in
the osh-worker daemon, it is better to let administrator restart it
manually when they are ready for that (e.g. after setting `max_load`
to zero and waiting for the already running tasks to finish).

Resolves: https://issues.redhat.com/browse/OSH-221
Closes: https://github.com/openscanhub/openscanhub/pull/217